Virgin_Mobile3-20240317T174127Z-001/Virgin_Mobile3/io/headspin/vm_pom_pytest/requirements.txt
pytest== 8.0.0
appium-python-client==2.6.0
selenium==4.16.0
requests==2.28.1